confmod/pyproject.toml
[tool.poetry]
package-mode = true
name = "libconfmod"
version = "0.0.1"
description = ""
authors = ["Maximilian Sudmann <maximilian.sudmann@rwh-aachen.de>"]
readme = "README.md"
exclude = ["domain", "scripts"]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
numpy = "^1.26.4"
pandas = "^2.2.2"
matplotlib = "3.7"
ipython = "^8.23.0"
notebook = "^7.1.3"
tikzplotlib = "^0.10.1"
adbc-driver-manager = "^0.11.0"
adbc-driver-postgresql = "^0.11.0"
pyarrow = "^16.0.0"
psycopg = "^3.1.18"
sqlalchemy = "^2.0.29"
psycopg2 = "^2.9.9"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.1"
click = "^8.1.7"
pydantic = "^2.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
